Fix compiler warnings on SunOS
utils.c: In function 'SendHotplugSignal':
utils.c:97:3: warning: format '%d' expects argument of type 'int', but argument 6 has type 'pid_t' [-Wformat=]
   Log2(PCSC_LOG_INFO, "Send hotplug signal to pcscd (pid=%d)", pid);
   ^
utils.c:100:4: warning: format '%d' expects argument of type 'int', but argument 6 has type 'pid_t' [-Wformat=]
    Log3(PCSC_LOG_CRITICAL, "Can't signal pcscd (pid=%d): %s",
    ^

+++ src/utils.c
@@ -94,11 +94,12 @@ int SendHotplugSignal(void)
 
 	if (pid != -1)
 	{
-		Log2(PCSC_LOG_INFO, "Send hotplug signal to pcscd (pid=%d)", pid);
+		Log2(PCSC_LOG_INFO, "Send hotplug signal to pcscd (pid=%ld)",
+			(long)pid);
 		if (kill(pid, SIGUSR1) < 0)
 		{
-			Log3(PCSC_LOG_CRITICAL, "Can't signal pcscd (pid=%d): %s",
-				pid, strerror(errno));
+			Log3(PCSC_LOG_CRITICAL, "Can't signal pcscd (pid=%ld): %s",
+				(long)pid, strerror(errno));
 			return EXIT_FAILURE ;
 		}
 		(void)SYS_Sleep(1);
